chore: rename useEnoughAllowance
